+++ babel.config.js
@@ -1,0 +1,43 @@
+module.exports = {
+  only: [
+    'src',
+    'styleguide',
+  ],
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        modules: false,
+      },
+    ],
+    '@babel/preset-react',
+  ],
+  plugins: [
+    'styled-components',
+    'babel-plugin-macros',
+    '@babel/plugin-syntax-dynamic-import',
+    '@babel/plugin-syntax-import-meta',
+    '@babel/plugin-proposal-class-properties',
+    '@babel/plugin-proposal-json-strings',
+    [
+      '@babel/plugin-proposal-decorators',
+      {
+        legacy: true,
+      },
+    ],
+  ],
+  env: {
+    production: {
+      plugins: [
+        'add-module-exports',
+        '@babel/plugin-transform-modules-commonjs',
+      ],
+    },
+    test: {
+      plugins: [
+        '@babel/plugin-transform-modules-commonjs',
+        'dynamic-import-node',
+      ],
+    },
+  },
+};