+++ src/theme/index.js
@@ -1,6 +1,7 @@
-import Color from 'color';
+import Color from '@bootstrap-styled/color';
 import unitUtils from '@bootstrap-styled/utils/lib/unitUtils';
 import { allowFalseValue, assertAscending, assertStartAtZero } from './utils';
+import { linearGradientRe } from '../utils/regex';
 
 const { detectUnit, rmUnit } = unitUtils;
 
@@ -214,14 +215,7 @@ function makeOriginal(userTheme = {}) {
   // Set the number of columns and specify the width of the gutters.
 
   v['$grid-columns'] = u['$grid-columns'] || '12';
-  v['$grid-gutter-width-base'] = u['$grid-gutter-width-base'] || '30px';
-  v['$grid-gutter-widths'] = u['$grid-gutter-widths'] || {
-    xs: v['$grid-gutter-width-base'],
-    sm: v['$grid-gutter-width-base'],
-    md: v['$grid-gutter-width-base'],
-    lg: v['$grid-gutter-width-base'],
-    xl: v['$grid-gutter-width-base'],
-  };
+  v['$grid-gutter-width'] = u['$grid-gutter-width'] || '30px';
 
   // Fonts
   //
@@ -348,31 +342,45 @@ function makeOriginal(userTheme = {}) {
   v['$btn-font-weight'] = u['$btn-font-weight'] || v['$font-weight-normal'];
   v['$btn-box-shadow'] = u['$btn-box-shadow'] || `inset 0 1px 0 ${Color(v['$white']).alpha(0.15).toString()}, 0 1px 1px ${Color(v['$black']).alpha(0.075).toString()}`;
   v['$btn-focus-box-shadow'] = u['$btn-focus-box-shadow'] || `0 0 0 2px ${Color(v['$brand-primary']).alpha(0.25).toString()}`;
+  v['$btn-disabled-opacity'] = u['$btn-disabled-opacity'] || '.65';
   v['$btn-active-box-shadow'] = u['$btn-active-box-shadow'] || `inset 0 3px 5px ${Color(v['$black']).alpha(0.125).toString()}`;
 
   v['$btn-primary-color'] = u['$btn-primary-color'] || v['$white'];
   v['$btn-primary-bg'] = u['$btn-primary-bg'] || v['$brand-primary'];
-  v['$btn-primary-border'] = u['$btn-primary-border'] || v['$btn-primary-bg'];
 
   v['$btn-secondary-color'] = u['$btn-secondary-color'] || v['$gray-dark'];
   v['$btn-secondary-bg'] = u['$btn-secondary-bg'] || v['$white'];
-  v['$btn-secondary-border'] = u['$btn-secondary-border'] || '#ccc';
 
   v['$btn-info-color'] = u['$btn-info-color'] || v['$white'];
   v['$btn-info-bg'] = u['$btn-info-bg'] || v['$brand-info'];
-  v['$btn-info-border'] = u['$btn-info-border'] || v['$btn-info-bg'];
 
   v['$btn-success-color'] = u['$btn-success-color'] || v['$white'];
   v['$btn-success-bg'] = u['$btn-success-bg'] || v['$brand-success'];
-  v['$btn-success-border'] = u['$btn-success-border'] || v['$btn-success-bg'];
 
   v['$btn-warning-color'] = u['$btn-warning-color'] || v['$white'];
   v['$btn-warning-bg'] = u['$btn-warning-bg'] || v['$brand-warning'];
-  v['$btn-warning-border'] = u['$btn-warning-border'] || v['$btn-warning-bg'];
 
   v['$btn-danger-color'] = u['$btn-danger-color'] || v['$white'];
   v['$btn-danger-bg'] = u['$btn-danger-bg'] || v['$brand-danger'];
-  v['$btn-danger-border'] = u['$btn-danger-border'] || v['$btn-danger-bg'];
+
+  /* this improvement will solve the linear-gradient when used for the background */
+  [
+    'primary',
+    'secondary',
+    'info',
+    'success',
+    'warning',
+    'danger',
+  ].forEach((type) => {
+    if (v[`$btn-${type}-bg`].includes('linear-gradient')) {
+      v[`$btn-${type}-border`] = v[`$btn-${type}-bg`].match(linearGradientRe)[1]; // eslint-disable-line prefer-destructuring
+    } else if (type === 'secondary') {
+      // secondary is having a white background, they use by default #ccc to make it look like a button
+      v['$btn-secondary-border'] = u['$btn-secondary-border'] || '#ccc';
+    } else {
+      v[`$btn-${type}-border`] = u[`$btn-${type}-border`] || v[`$btn-${type}-bg`];
+    }
+  });
 
   v['$btn-link-disabled-color'] = u['$btn-link-disabled-color'] || v['$gray-light'];
 
@@ -702,8 +710,9 @@ function makeOriginal(userTheme = {}) {
 
   v['$card-img-overlay-padding'] = u['$card-img-overlay-padding'] || '1.25rem';
 
-  detectedUnit = detectUnit(v['$grid-gutter-width-base']);
-  v['$card-deck-margin'] = u['$card-deck-margin'] || (rmUnit(v['$grid-gutter-width-base'], detectedUnit) / 2) + detectedUnit;
+  detectedUnit = detectUnit(v['$grid-gutter-width']);
+  v['$card-group-margin'] = u['$card-group-margin'] || (rmUnit(v['$grid-gutter-width'], detectedUnit) / 2) + detectedUnit;
+  v['$card-deck-margin'] = u['$card-deck-margin'] || v['$card-group-margin'];
 
   v['$card-columns-count-md'] = u['$card-columns-count-md'] || '2';
   v['$card-columns-gap-md'] = u['$card-columns-gap-md'] || '1rem';
@@ -931,6 +940,4 @@ function makeOriginal(userTheme = {}) {
 
 export const makeTheme = makeOriginal;
 
-const theme = makeOriginal();
-
-export default theme;
+export default makeOriginal();