+++ src/utils/index.js
@@ -38,7 +38,7 @@ import deepMerge from 'lodash.merge';
  * });
  */
 export const toMakeTheme = (theme) => (userTheme) => { // eslint-disable-line arrow-body-style
-  return deepMerge(theme, userTheme);
+  return deepMerge({}, theme, userTheme);
 };
 
 function makeThemeFromList(list, theme) {
@@ -70,7 +70,7 @@ export default function createMakeTheme(list) {
  *
  */
 export function makeScopedTheme(scopeName, userTheme = { [scopeName]: {} }) {
-  if (scopeName === undefined || typeof scopeName !== 'string') return console.warn('You may have forgotten to set the scope name in the makeScopedTheme function.');
+  if (scopeName === undefined || typeof scopeName !== 'string') return console.warn('You may have forgotten to set the scope name in the makeScopedTheme function.'); // eslint-disable-line no-console
   const newTheme = { [scopeName]: {} };
   const v = newTheme[scopeName];
   const u = userTheme[scopeName] || {};